refactor: GameSettingInput 분리
